+++ sdk/ai/ai-projects/samples/v1-beta/javascript/evaluations/evaluationBasics.js
@@ -43,7 +43,7 @@ async function main() {
     },
   };
 
-  const evalResp = await project.evaluations.createRun(newEvaluation);
+  const evalResp = await project.evaluations.create(newEvaluation);
   console.log("Create a new evaluation:", JSON.stringify(evalResp, null, 2));
   // get the evaluation by ID
   const eval2 = await project.evaluations.get(evalResp.name);
@@ -57,7 +57,6 @@ async function main() {
   }
   console.log("List of evaluation display names:", evaluationNames);
 
-  // This is temporary, as interface recommend the name of the evaluation
   const name = evaluations[0].name;
   const evaluation = await project.evaluations.get(name);
   console.log("Get an evaluation by ID:", JSON.stringify(evaluation, null, 2));